[#154747783] Renamed sqolToApiMappers
