+++ app/javascript/components/mappers/soqlToApiFields.js
